requirements.txt
numpy==1.23.3
scikit-learn==1.1.2
PyYAML==6.0
openpyxl==3.0.10
pandas==1.5.0
scipy==1.9.2